Move documentation back to azuracast.com repo
Rationale: With automatic updates being sent out with every commit to this repository, there is no need for station maintainers to update in the event of changes to the published documentation. The addition of Netlify to this repo has also added a lot of unnecessary Pull Request checks for changes that have nothing to do with the documentation. Using the separate repository, we can create as many Markdown files (both on and off of the public site) as we want without prompting station owners for updates.