Fixes for file playtime detection
StationMediaRepository.php:  sometimes getID3 will report an error in a file
due to a slightly mangled header but can still get an accurate playtime length
and some other info. In the loadFromFile function removed the condition that
there are no errors prior to checking playtime so that we can still get that info,
and instead sent any detected errors to the php log.

Flow.php: In _createFileFromChunks function changed final filename to use actual
filename uploaded instead of putting _complete at the end so that getID3 can
use the file extension as a fallback identifier if the standard mp3 header is
slightly mangled. That fallback occurs in GetFileFormat() in
/var/azuracast/www/vendor/james-heinrich/getid3/getid3/getid3.php