TuneIn Webhook actually send the album instead of the artist twice. (#2863)
The TuneIn Webhook currently sends the artist name as the album which is not correct obviously. Because of this, the TuneIn interface would show Artist - Song (Artist) instead of Artist - Song (Album).

This $np->now_playing->song->album should exist as it is in the Now Playing API. When empty it should be ignored and TuneIn wll deals with it.